feat: store embeddings as f64's for compatibility with analyses
this will let us reuse the reclustering code for analyses for embeddings

+++ storage/src/db/schemas/analysis.rs
@@ -26,7 +26,7 @@ pub struct Analysis {
     pub features: [f64; NUMBER_FEATURES],
 
     /// The [`Song`]'s embedding vector.
-    pub embedding: [f32; mecomp_analysis::DIM_EMBEDDING],
+    pub embedding: [f64; mecomp_analysis::DIM_EMBEDDING],
 }
 
 #[cfg(feature = "db")]
